requirements.txt
asgiref==3.5.2
async-generator==1.10      
attrs==21.4.0
autopep8==1.6.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12 
commonmark==0.9.1
cryptography==37.0.2       
Django==4.0.5
django-ckeditor==6.4.2     
django-crispy-forms==1.14.0
django-emoji==2.2.2        
django-emoji-picker==0.0.6 
django-js-asset==2.0.0     
h11==0.13.0
idna==3.3
MouseInfo==0.1.3
outcome==1.2.0
Pillow==9.1.1
PyAutoGUI==0.9.53
PyAutoIt==0.6.5
pycodestyle==2.8.0
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.12.0
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dotenv==0.20.0
pytweening==1.0.4
requests==2.28.0
rich==12.4.4
selenium==4.2.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
toml==0.10.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.9
webdriver-manager==3.7.0
wsproto==1.1.0